Created messaging class for issue reports and volunteer requests.
