+++ Seattle Trails/AppExtensions.swift
@@ -102,6 +102,7 @@ extension UIImagePickerController {
     func presentImagePickerWithSourceTypeForViewController<VC: UIViewController where VC: UIImagePickerControllerDelegate, VC: GetsImageToShare>(sender: VC, sourceType: UIImagePickerControllerSourceType)
     {
         sender.imagePicker.sourceType = sourceType
+        
         dispatch_async(dispatch_get_main_queue())
         {
             sender.presentViewController(sender.imagePicker, animated: true, completion: nil)
